Add Password locking things, ?unlock feature, make buttons easier to press
I added a Password option to the settings it works on the index page the key is alt+l to lock. I also made the buttons easier to click (I forgot to make an or statement for the other value lol). I'll add the password lock to all pages in the next commit. I also started adding things so we can get rammerhead working.